% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cLambda12_frailty_lk.R
\name{cLambda12_frailty_lk}
\alias{cLambda12_frailty_lk}
\title{likelihood function}
\usage{
cLambda12_frailty_lk(x, xdata1, omega1, shape12, scale12, c12, beta12_1)
}
\arguments{
\item{x}{xdata1
omega1
shape12
scale12
c12
beta12_1}
}
\value{
likelihood
}
\description{
evaluate cumulative hazard
}
\examples{
example(cLambda12_frailty_lk(x = 1, xdata1 = 0, omega1 = 1, shape12 = 1, scale12 = 1, c12 = 1, beta12_1 = 0))
}
