% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cLambda13_frailty_lk.R
\name{cLambda13_frailty_lk}
\alias{cLambda13_frailty_lk}
\title{likelihood function}
\usage{
cLambda13_frailty_lk(x, xdata1, omega2, scale13, shape13, c13, beta13_1)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(cLambda13_frailty_lk(x = 1, xdata1 = 0, omega2 = 1, shape13 = 1, scale13 = 1, c13 = 1, beta13_1 = 0))
}
