% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_results.R
\name{final_results}
\alias{final_results}
\title{final results function}
\usage{
final_results(params_matrix, write)
}
\arguments{
\item{params_matrix}{parameter matrix of results}

\item{write}{logical value if results should be written to a file}
}
\value{
table
}
\description{
tabulate results from mcmc
}
\examples{
example(final_results(params_matrix = params_matrix, write = T))
}
