% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like12_omega_i.R
\name{like12_omega_i}
\alias{like12_omega_i}
\title{likelihood function for frailty}
\usage{
like12_omega_i(omega12_z0, i, scale12_0, dat0, shape12_0, c12_0)
}
\arguments{
\item{omega12_z0:}{frailty term
i: index for individual in dataset
dat0: dataset
shape12_0: shape for Weibull distribution
scale12_0: scale for Weibull distribution
c12_0: frailty coefficient}
}
\value{
likelihood
}
\description{
evaluate likelihood for frailty for 1-2 transition
}
\examples{
example(like12_omega1_i(omega12_z0 = 1, i = 1, scale12_0 = 1, dat0 = dat0, shape12_0 = 1, c12_0 = 1))
}
