% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like12_shape.R
\name{like12_shape}
\alias{like12_shape}
\title{likelihood function}
\usage{
like12_shape(shape12_0, scale12_0, dat0, c12_0, omega12_z0)
}
\arguments{
\item{weibull}{model parameters
omega12_z0: frailty term
shape12_0: shape of Weibull distribution
scale12_0: scale of Weibull distribution
c12_0: frailty coefficient
dat0: data}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(like12_shape(shape12_0 = 1, scale12_0 = 1, dat0 = 1, c12_0 = 1, omega12_z0 = 1))
}
