% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like13_omega1_i.R
\name{like13_omega1_i}
\alias{like13_omega1_i}
\title{likelihood function for frailty}
\usage{
like13_omega1_i(
  omega13_z1,
  i,
  scale13_1,
  dat1,
  shape13_1,
  c13_1,
  scale23_1,
  shape23_1,
  c23_1,
  theta23_1
)
}
\arguments{
\item{omega13_z1:}{frailty term
i: index for individual in dataset
dat1: dataset
shape13_1: shape for Weibull distribution
scale13_1: scale for Weibull distribution
c13_1: frailty coefficient
shape23_1: shape for Weibull distribution
scale23_1: scale for Weibull distribution
c23_1: frailty coefficient
theta23_1: theta parameter for 2-3 transition}
}
\value{
likelihood
}
\description{
evaluate likelihood for frailty for 1-3 transition
}
\examples{
example(like13_omega1_i(omega13_z1 = 1, i = 1, scale13_1 = 1, dat1 = dat1, shape13_1 = 1, c13_1 = 1, scale23_1 = 1, shape23_1 = 1, c23_1 = 1, theta23_1 = 0))
}
