% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like13_omega_i.R
\name{like13_omega_i}
\alias{like13_omega_i}
\title{likelihood function}
\usage{
like13_omega_i(
  omega13_z0,
  i,
  scale13_0,
  dat0,
  shape13_0,
  c13_0,
  scale23_0,
  shape23_0,
  c23_0,
  theta23_0
)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(like13_omega_i(omega13_z0 = 1, i = 1, scale13_0 = 1, dat1 = dat1, shape13_0 = 1, c13_0 = 1, scale23_0 = 1, shape23_0 = 1, c23_0 = 1, theta23_0 = 0))
}
