% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like13_shape.R
\name{like13_shape}
\alias{like13_shape}
\title{likelihood function}
\usage{
like13_shape(shape13_0, dat0, c13_0, scale13_0, omega13_z0)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
likelihood for given parameter set

}
