% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like13_shape1.R
\name{like13_shape1}
\alias{like13_shape1}
\title{likelihood function}
\usage{
like13_shape1(shape13_1, dat1, c13_1, scale13_1, omega13_z1)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
likelihood for given parameter set

}
