% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like23_shape.R
\name{like23_shape}
\alias{like23_shape}
\title{likelihood function}
\usage{
like23_shape(shape23_0, dat0, c23_0, scale23_0, theta23_0, omega13_z0)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(like23_shape(shape23_0 = 1, dat0 = 1, c23_0 = 1, scale23_0 = 1, theta23_0 = 1, omega13_z0 = 1))
}
