% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like23_shape1.R
\name{like23_shape1}
\alias{like23_shape1}
\title{likelihood function}
\usage{
like23_shape1(shape23_1, dat1, c23_1, scale23_1, theta23_1, omega13_z1)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(like23_shape1(shape23_1 = 1, dat1 = 1, c23_1 = 1, scale23_1 = 1, theta23_1 = 1, omega13_z1 = 1))
}
