% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cLambda12_frailty_lk.R
\name{cLambda12_frailty_lk}
\alias{cLambda12_frailty_lk}
\title{cumulative hazard function}
\usage{
cLambda12_frailty_lk(x, xdata1, omega1, shape12, scale12, c12, beta12_1)
}
\arguments{
\item{x}{value at which to evaluate cumulative hazard}

\item{xdata1}{covariate data}

\item{omega1}{frailty term}

\item{shape12}{shape of Weibull distribution}

\item{scale12}{scale of Weibull distribution}

\item{c12}{frailty coefficient}

\item{beta12_1}{covariate coefficient}
}
\value{
cumulative hazard
}
\description{
evaluate cumulative hazard for the 1-2 transition
}
\examples{
example(cLambda12_frailty_lk(x = 1, xdata1 = 0, omega1 = 1, 
shape12 = 1, scale12 = 1, c12 = 1, beta12_1 = 0))
}
