% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cLambda12_frailty_lk.R
\name{cLambda12_frailty_lk}
\alias{cLambda12_frailty_lk}
\title{likelihood function}
\usage{
cLambda12_frailty_lk(x, xdata1, omega1, shape12, scale12, c12, beta12_1)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
likelihood for given parameter set

}
