% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cLambda13_frailty_lk.R
\name{cLambda13_frailty_lk}
\alias{cLambda13_frailty_lk}
\title{cumulative hazard function}
\usage{
cLambda13_frailty_lk(x, xdata1, omega2, scale13, shape13, c13, beta13_1)
}
\arguments{
\item{x}{value at which to evaluate cumulative hazard}

\item{xdata1}{covariate data}

\item{omega2}{frailty term}

\item{scale13}{scale of Weibull distribution}

\item{shape13}{shape of Weibull distribution}

\item{c13}{frailty coefficient}

\item{beta13_1}{covariate coefficient}
}
\value{
cumulative hazard
}
\description{
evaluate cumulative hazard for the 1-3 transition
}
\examples{
example(cLambda13_frailty_lk(x = 1, xdata1 = 0, omega2 = 1, 
shape13 = 1, scale13 = 1, c13 = 1, beta13_1 = 0))
}
