% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cLambda23_frailty_lk.R
\name{cLambda23_frailty_lk}
\alias{cLambda23_frailty_lk}
\title{likelihood function}
\usage{
cLambda23_frailty_lk(
  x,
  xdata1,
  v_predict,
  omega2,
  shape23,
  scale23,
  omega1,
  theta23,
  c23,
  beta23_1
)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
likelihood for given parameter set

}
