% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda12_frailty.R
\name{lambda12_frailty}
\alias{lambda12_frailty}
\title{likelihood function}
\usage{
lambda12_frailty(x, xdata1, omega1, scale12, shape12, c12, beta12_1)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(lambda12_frailty(x = 1, xdata1 = 0, omega1 = 1, shape12 = 1, scale12 = 1, c12 = 1, beta12_1 = 0))
}
