% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like12_omega_i.R
\name{like12_omega_i}
\alias{like12_omega_i}
\title{likelihood function}
\usage{
like12_omega_i(omega12_z0, i, scale12_0, z, dat0, shape12_0, c12_0)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
likelihood for given parameter set

}
