% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like12_shape.R
\name{like12_shape}
\alias{like12_shape}
\title{likelihood function}
\usage{
like12_shape(shape12_0, scale12_0, dat0, c12_0, omega12_z0)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
likelihood for given parameter set

}
