% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like12_shape1.R
\name{like12_shape1}
\alias{like12_shape1}
\title{likelihood function}
\usage{
like12_shape1(shape12_1, scale12_1, dat1, c12_1, omega12_z1)
}
\arguments{
\item{shape12_1}{shape of Weibull distribution}

\item{scale12_1}{scale of Weibull distribution}

\item{dat1}{data}

\item{c12_1}{frailty coefficient}

\item{omega12_z1}{frailty term}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(like12_shape1(shape12_1 = 1, scale12_1 = 1, dat1 = 1,
c12_1 = 1, omega12_z1 = 1))
}
