% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like13_omega_i.R
\name{like13_omega_i}
\alias{like13_omega_i}
\title{likelihood function for frailty}
\usage{
like13_omega_i(
  omega13_z0,
  i,
  scale13_0,
  dat0,
  shape13_0,
  c13_0,
  scale23_0,
  shape23_0,
  c23_0,
  theta23_0
)
}
\arguments{
\item{omega13_z0:}{frailty term
i: index for individual in dataset
dat0: dataset
shape13_0: shape for Weibull distribution
scale13_0: scale for Weibull distribution
c13_1: frailty coefficient
shape23_0: shape for Weibull distribution
scale23_0: scale for Weibull distribution
c23_0: frailty coefficient
theta23_0: theta parameter for 2-3 transition}
}
\value{
likelihood
}
\description{
evaluate likelihood for frailty for 1-3 transition
}
\examples{
example(like13_omega_i(omega13_z0 = 1, i = 1, scale13_0 = 1, dat1 = dat1, shape13_0 = 1, c13_0 = 1, scale23_0 = 1, shape23_0 = 1, c23_0 = 1, theta23_0 = 0))
}
