% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like13_shape.R
\name{like13_shape}
\alias{like13_shape}
\title{likelihood function}
\usage{
like13_shape(shape13_0, dat0, c13_0, scale13_0, omega13_z0)
}
\arguments{
\item{omega13_z0:}{frailty term
shape13_0: shape of Weibull distribution
scale13_0: scale of Weibull distribution
c13_0: frailty coefficient
dat0: data}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(like13_shape(shape13_0 = 1, dat0 = 1, c13_0 = 1, scale13_0 = 1, omega13_z0 = 1))
}
