% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like13_shape1.R
\name{like13_shape1}
\alias{like13_shape1}
\title{likelihood function}
\usage{
like13_shape1(shape13_1, dat1, c13_1, scale13_1, omega13_z1)
}
\arguments{
\item{shape13_1}{shape of Weibull distribution}

\item{dat1}{data}

\item{c13_1}{frailty coefficient}

\item{scale13_1}{scale of Weibull distribution}

\item{omega13_z1}{frailty term}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(like13_shape1(shape13_1 = 1, dat1 = 1, c13_1 = 1, 
scale13_1 = 1, omega13_z1 = 1))
}
