% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like23_theta.R
\name{like23_theta}
\alias{like23_theta}
\title{likelihood function}
\usage{
like23_theta(theta23_0, dat0, c23_0, shape23_0, scale23_0, omega13_z0)
}
\arguments{
\item{weibull}{model parameters}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
likelihood for given parameter set

}
