% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like23_theta.R
\name{like23_theta}
\alias{like23_theta}
\title{likelihood function}
\usage{
like23_theta(theta23_0, dat0, c23_0, shape23_0, scale23_0, omega13_z0)
}
\arguments{
\item{weibull}{model parameters
omega13_z0: frailty term
shape23_0: shape of Weibull distribution
scale23_0: scale of Weibull distribution
c23_0: frailty coefficient
theta23_0: covariate coefficient
dat0: data}
}
\value{
likelihood
}
\description{
evaluate likelihood ...
}
\examples{
example(like23_theta(theta23_0 = 1, dat0 = 1, c23_0 = 1, shape23_0 = 1, scale23_0 = 1, omega13_z0 = 1))
}
