% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{plot results from mcmc}
\usage{
plot_results(params_matrix, write)
}
\arguments{
\item{results}{from mcmc}
}
\value{
cep curve
}
\description{
plot the results in a cep curve
}
\examples{
example(plot_results(params_matrix = params_matrix, write = F))
}
