% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{run the simulation}
\usage{
run_sim(SIM, rhos, rhot, frailtysd, params_list)
}
\arguments{
\item{SIM:}{number of iterations of mcmc to run
rhos: correlation parameter
rhot: correlation parameter
params_list: list of true generating parameters}
}
\value{
simulation results
}
\description{
run mcmc simulation
}
\examples{
example(run_sim(SIM = 3000, rhos = 0.5, rhot = 0.5, frailtysd = 1, params_list = params_list))
}
