% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{simulate data for simulations}
\usage{
sim_data(n, array_id, scenario, effecttheta, frailtysd, rhot, rhos, rhost)
}
\arguments{
\item{generating}{parameters}
}
\value{
data
}
\description{
simulate data
}
\examples{
create data

}
