% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_cep.R
\name{true_cep}
\alias{true_cep}
\title{create true cep curve}
\usage{
true_cep(dat0, dat1, write, params_list, plotwrite)
}
\arguments{
\item{dat0:}{data for placebo arm
dat1: data for treatment arm
write: logical value if results should be written to a file
params_list: true values of parameters
plotwrite: logical value if plot should be saved to a file}
}
\value{
true values of CEP curve
}
\description{
create cep true based on true values
}
\examples{
example(true_cep(dat0 = dat0, dat1 = dat1, write = T, params_list = params_list, plotwrite = T))
}
