% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_cep.R
\name{true_cep}
\alias{true_cep}
\title{create true cep curve}
\usage{
true_cep(dat0, dat1, write, params_list, plotwrite, fignum, tau_s, tau_t)

true_cep(dat0, dat1, write, params_list, plotwrite, fignum, tau_s, tau_t)
}
\arguments{
\item{dat0}{data for placebo arm}

\item{dat1}{data for treatment arm}

\item{write}{logical value if results should be written to a file}

\item{params_list}{true values of parameters}

\item{plotwrite}{logical value if plot should be saved to a file}

\item{fignum}{figure number for saving}

\item{tau_s}{time of evaluation for S}

\item{tau_t}{time of evaluation for T}
}
\value{
true values of CEP curve

true values of CEP curve
}
\description{
create cep true based on true values

create cep true based on true values
}
\examples{
example(true_cep(dat0 = dat0, dat1 = dat1, write = T, params_list = 
params_list, plotwrite = T))
example(true_cep(dat0 = dat0, dat1 = dat1, write = T, params_list = 
params_list, plotwrite = T))
}
